+++ sample/src/main/java/com/davidecirillo/multichoicesample/sampleToolbar/SampleToolbarActivity.java
@@ -2,6 +2,7 @@ package com.davidecirillo.multichoicesample.sampleToolbar;
 
 import android.content.Intent;
 import android.os.Bundle;
+import android.support.annotation.IntDef;
 import android.support.annotation.Nullable;
 import android.support.v7.widget.GridLayoutManager;
 import android.support.v7.widget.LinearLayoutManager;
@@ -16,6 +17,8 @@ import com.davidecirillo.multichoicesample.BaseActivity;
 import com.davidecirillo.multichoicesample.R;
 import com.davidecirillo.multichoicesample.ResultActivity;
 
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -29,12 +32,27 @@ import rx.schedulers.Schedulers;
 
 public class SampleToolbarActivity extends BaseActivity {
 
+    private static final int DEFAULT_QUANTITY_MODE = QuantityMode.STRING;
+
+    @IntDef({
+            QuantityMode.NONE,
+            QuantityMode.STRING,
+            QuantityMode.PLURALS,
+    })
+    @Retention(RetentionPolicy.SOURCE)
+    public @interface QuantityMode {
+        int NONE = 0;
+        int STRING = 1;
+        int PLURALS = 2;
+    }
+
     public static final String SELECTED_ITEMS = "selectedItems";
     @BindView(R.id.multiChoiceRecyclerView)
     RecyclerView mMultiChoiceRecyclerView;
 
     private ArrayList<String> stringList;
     private MySampleToolbarAdapter mMySampleToolbarAdapter;
+    private @SampleToolbarActivity.QuantityMode int quantityMode = DEFAULT_QUANTITY_MODE;
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
@@ -88,17 +106,33 @@ public class SampleToolbarActivity extends BaseActivity {
     private void setUpAdapter(List<String> strings) {
         stringList = new ArrayList<>(strings);
 
-        MultiChoiceToolbar multiChoiceToolbar =
-                new MultiChoiceToolbar.Builder(SampleToolbarActivity.this, toolbar)
-                        .setTitles(toolbarTitle(), "item selected")
-                        .setMultiChoiceColours(R.color.colorPrimaryMulti, R.color.colorPrimaryDarkMulti)
-                        .setDefaultIcon(R.drawable.ic_arrow_back_white_24dp, new View.OnClickListener() {
-                            @Override
-                            public void onClick(View view) {
-                                onBackPressed();
-                            }
-                        })
-                        .build();
+        MultiChoiceToolbar.Builder builder = new MultiChoiceToolbar.Builder(SampleToolbarActivity.this, toolbar)
+                .setMultiChoiceColours(R.color.colorPrimaryMulti, R.color.colorPrimaryDarkMulti)
+                .setDefaultIcon(R.drawable.ic_arrow_back_white_24dp, new View.OnClickListener() {
+                    @Override
+                    public void onClick(View view) {
+                        onBackPressed();
+                    }
+                });
+
+        switch (quantityMode) {
+            case QuantityMode.NONE:
+                /*
+                 * You may use one of the setTitles()-methods
+                 * or none (it will then show the count of selected items instead).
+                 */
+                break;
+
+            case QuantityMode.STRING:
+                builder.setTitles(toolbarTitle(), "item selected");
+                break;
+
+            case QuantityMode.PLURALS:
+                builder.setTitles(toolbarTitle(), R.plurals.numberOfSelectedItems);
+                break;
+        }
+
+        MultiChoiceToolbar multiChoiceToolbar = builder.build();
 
         mMySampleToolbarAdapter = new MySampleToolbarAdapter(stringList, getApplicationContext());
         mMySampleToolbarAdapter.setMultiChoiceToolbar(multiChoiceToolbar);
@@ -126,26 +160,36 @@ public class SampleToolbarActivity extends BaseActivity {
 
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
-
         if (mMySampleToolbarAdapter != null) {
             switch (item.getItemId()) {
                 case R.id.select_all:
                     mMySampleToolbarAdapter.selectAll();
                     return true;
+
                 case R.id.deselect_all:
                     mMySampleToolbarAdapter.deselectAll();
                     return true;
+
                 case R.id.select_3:
                     mMySampleToolbarAdapter.select(2);
                     return true;
-                case R.id.single_click_mode:
 
+                case R.id.single_click_mode:
                     mMySampleToolbarAdapter.setSingleClickMode(!mMySampleToolbarAdapter.isInSingleClickMode());
-                    Toast.makeText(getApplicationContext(), "Always Single Click Mode [" + mMySampleToolbarAdapter.isInSingleClickMode() + "]", Toast
-                            .LENGTH_SHORT).show();
+                    Toast.makeText(getApplicationContext(),
+                            "Always Single Click Mode [" + mMySampleToolbarAdapter.isInSingleClickMode() + "]",
+                            Toast.LENGTH_SHORT).show();
+
+                    return true;
+
+                case R.id.plural_mode:
+                    setQuantityMode(QuantityMode.PLURALS);
+
+                    Toast.makeText(getApplicationContext(), "set toolbar to use QuantityMode.PLURALS", Toast.LENGTH_SHORT).show();
                     return true;
             }
         }
+
         return false;
     }
 
@@ -157,5 +201,17 @@ public class SampleToolbarActivity extends BaseActivity {
     @Override
     protected String toolbarTitle() {
         return getString(R.string.toolbar_controls);
+    }
+
+    @QuantityMode
+    public int getQuantityMode() {
+        return quantityMode;
+    }
+
+    public void setQuantityMode(@QuantityMode int quantityMode) {
+        mMySampleToolbarAdapter.deselectAll();
+
+        this.quantityMode = quantityMode;
+        setUpMultiChoiceRecyclerView();
     }
 }